+++ examples/sorting/main.rs
@@ -33,26 +33,26 @@ fn main() {
 
     rgsl::sort::select::sort_smallest(&mut small, k, &x, 1);
     println!("{} smallest values from {}", k, n);
-    for tmp in range(0, k as uint) {
+    for tmp in range(0, k as usize) {
         println!("{}: {}", k, small[tmp]);
     }
 
     rgsl::sort::select::sort_largest(&mut small, k, &x, 1);
     println!("\n{} largest values from {}", k, n);
-    for tmp in range(0, k as uint) {
+    for tmp in range(0, k as usize) {
         println!("{}: {}", k, small[tmp]);
     }
 
     small.swap(2, 3);
     rgsl::sort::objects::heapsort_index(&mut p, &small, compare_func);
     println!("\nheapsort_index :", );
-    for tmp in range(0, k as uint) {
+    for tmp in range(0, k as usize) {
         println!("{}: {}", k, p[tmp]);
     }
 
     rgsl::sort::objects::heapsort(&mut small, compare_func);
     println!("\nheapsort :", );
-    for tmp in range(0, k as uint) {
+    for tmp in range(0, k as usize) {
         println!("{}: {}", k, small[tmp]);
     }
 }
